Clean up project structure
Clean up project strucutre.
